read-tree checkpoint

+++ src/cmd/read_tree.rs
@@ -1,15 +1,17 @@
+use crate::cmd::cat_file;
 use std::collections::HashSet;
-use std::fs::{self, File};
-use std::io::Result;
+use std::fs;
+use std::io::{BufRead, BufReader, Error, ErrorKind, Result, Write};
 use std::path::{Path, PathBuf};
-
 pub fn main(hash: &str) -> Result<()> {
-    let mut path = PathBuf::from(".gyat/objects");
-    path.push(hash);
-    let file = File::open(path).expect("Invalid hash provided");
+    let path = PathBuf::from(".gyat/objects").join(hash);
+    if !Path::new(&path).exists() {
+        return Err(Error::new(ErrorKind::InvalidInput, "Invalid hash provided"));
+    }
 
     let ignored_files = gyat::get_ignored_file_list();
     delete_cwd(".", &ignored_files)?;
+    restore_cwd(".", hash)?;
 
     Ok(())
 }
@@ -32,4 +34,41 @@ fn delete_cwd<P: AsRef<Path>>(path: P, ignored_files: &HashSet<String>) -> Resul
 
     fs::remove_dir(path)?;
     Ok(())
+}
+
+fn restore_cwd<P: AsRef<Path>>(path: P, hash: &str) -> Result<()> {
+    let objpath = PathBuf::from(".gyat/objects").join(hash);
+    let file = fs::File::open(objpath)?;
+    let reader = BufReader::new(file);
+    let lines: Vec<String> = reader.lines().collect::<Result<_>>().unwrap_or_default();
+    let lines = parse_tree_file(lines);
+
+    for (filetype, hash, filename) in lines {
+        let path = PathBuf::from(path.as_ref()).join(filename);
+        if filetype == "tree" {
+            fs::create_dir(path.clone())?;
+            restore_cwd(path, &hash)?;
+        } else {
+            let contents = cat_file::cat_file(&hash)?;
+            let mut file = fs::File::create(path)?;
+            file.write_all(contents.as_bytes())?;
+            file.flush()?;
+        }
+    }
+
+    Ok(())
+}
+
+fn parse_tree_file(input: Vec<String>) -> Vec<(String, String, String)> {
+    input
+        .into_iter()
+        .map(|s| {
+            let mut parts = s.split_whitespace();
+            (
+                parts.next().unwrap().to_string(),
+                parts.next().unwrap().to_string(),
+                parts.next().unwrap().to_string(),
+            )
+        })
+        .collect()
 }
